separation of concerns and self-documenting code
